refactor: rename redux folder to store for aliases
