+++ src/components/SignUpForm/__test__/SignUpForm.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import SignUpForm from '../SignUpForm';
 import { Provider } from 'react-redux';
-import store from '../../../redux/store';
+import store from '../../../store/store';
 
 describe('SignUpForm tests', () => {
   it('SignUpForm renders without errors', () => {
