+++ src/pages/NotFoundPage/__test__/NotFoundPage.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import NotFoundPage from '../NotFoundPage';
 import { Provider } from 'react-redux';
 import RegionProvider from '../../../hoc/context/RegionContext/RegionProvider';
-import store from '../../../redux/store';
+import store from '../../../store/store';
 import { MemoryRouter } from 'react-router-dom';
 
 describe('NotFoundPage component', () => {
