+++ src/pages/WelcomePage/WelcomePage.tsx
@@ -7,7 +7,7 @@ import useRegion from '../../hook/useRegion';
 import { LOCALE_DATA } from '../../locales/constants/constants';
 import { Developers } from '../../components/Developers/Developers';
 import Loader from '../../components/Loader/Loader';
-import { PrimaryButton } from '../../components/UI/PrimaryButton/PrimaryButton';
+import { PrimaryButton } from 'components/UI/PrimaryButton/PrimaryButton';
 import CourseInfo from '../../components/CourseInfo/CourseInfo';
 
 function WelcomePage() {
