+++ src/store/slices/GraphQLSlice.ts
